Convert player to a class: Player.java
This updates all code that relies on Constant.Player.Human/Computer.

Player specific methods in GameAction.java are marked as deprecatred and redirected to Player.java.

TODO:
1. Incendiary Command needs some looking at.  I couldn't comletely fix it.
2) There is a TODO in the WheneverKeyword code with a very small, specific section commented out.

Everything else has been converted to the best of my knowledge.
	
Due to previously using Strings everywhere, this may introduce some bugs, especially NullPointerExceptions.

Everything uses the object except for the String[][] in Phase.java.  There is a hack in place to make it work, but that shuold probably be converted to a different data structure.
	
See the Developer's Forum thread for more complete info.